Alter the displayWeatherForecast function
Use the DocumentFragment interface so each iteration of the forEach loop instead of appending newly created nodes to the main container, it attaches them to the DocumentFragment (offscreen DOM tree). Since the document fragment is in memory and not part of the main DOM tree appending children to it does not cause page reflow (computation of element's position and geometry).